# specs/001-railway-rag-system/contracts/api.openapi.yaml
openapi: 3.0.3
info:
  title: BMS Agent - Railway Documentation RAG System API
  description: |
    API for semantic search and document management in railway technical documentation.

    **Features**:
    - Multi-format document upload (PDF, DOCX, PPTX, XLSX, CSV, TXT, MD)
    - Hierarchical chunking with quality validation
    - Semantic, hybrid, and hierarchical search modes
    - Department and railway-specific metadata filtering
    - Basic operational metrics and health checks

    **Authentication**: Optional API key via X-API-Key header (configurable)
  version: 1.0.0
  contact:
    name: BMS Agent Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.bms-agent.example.com
    description: Production server (if applicable)
security:
  - ApiKeyAuth: []
  - {}  # Allow anonymous access when API key is disabled
...
paths:
...
  /api/v1/search/semantic:
    post:
      tags:
        - Search
      summary: Semantic search
      description: |
        Perform pure semantic search using vector similarity.
        Returns top-k results ranked by cosine similarity.
      operationId: semanticSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              basic:
                summary: Basic query
                value:
                  query: "emergency braking procedures"
                  limit: 10
              filtered:
                summary: Department-filtered query
                value:
                  query: "maintenance schedules"
                  limit: 10
                  filter_department: "Maintenance"
                  filter_fleet_type: "EMU"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/search/hybrid:
    post:
      tags:
        - Search
      summary: Hybrid search (semantic + keyword)
      description: |
        Combine semantic and keyword search using Reciprocal Rank Fusion (RRF).
        Alpha parameter controls weight: 0=keyword only, 1=semantic only, 0.5=balanced.
      operationId: hybridSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/SearchRequest'
                - type: object
                  properties:
                    alpha:
                      type: number
                      format: float
                      minimum: 0.0
                      maximum: 1.0
                      default: 0.5
                      description: Weight for semantic vs keyword (0=keyword, 1=semantic)
                      example: 0.7
            examples:
              semantic_heavy:
                summary: Semantic-heavy hybrid search
                value:
                  query: "train safety protocols"
                  alpha: 0.8
                  limit: 10
              keyword_heavy:
                summary: Keyword-heavy for technical terms
                value:
                  query: "EMU-750 specifications"
                  alpha: 0.3
                  limit: 10
      responses:
        '200':
          description: Hybrid search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for authenticated access
  schemas:
...
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 3
          maxLength: 1000
          description: Natural language search query
          example: emergency braking procedures
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Number of results to return
        filter_department:
          type: string
          description: Filter by department
          example: Safety
        filter_category:
          type: string
          description: Filter by category
        filter_document_name:
          type: string
          description: Filter by document name
        filter_chapter_number:
          type: string
          description: Filter by chapter number
          example: "5.2.1"
...
    SearchResponse:
      type: object
      required:
        - query
        - results
        - total_results
        - search_time_ms
      properties:
        query:
          type: string
          description: Original search query
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total_results:
          type: integer
          description: Number of results returned
          example: 10
        search_time_ms:
          type: number
          format: float
          description: Search latency in milliseconds
          example: 145.2
        search_mode:
          type: string
          enum: [semantic, hybrid, hierarchical, railway_specific]
          description: Search mode used
        filters_applied:
          type: object
          description: Filters that were applied
          example:
            department: Safety
            fleet_type: EMU
    SearchResult:
      type: object
      required:
        - chunk_id
        - chunk_text
        - relevance_score
        - ranking_position
        - source_document
      properties:
        chunk_id:
          type: string
          format: uuid
        chunk_text:
          type: string
          description: Matched chunk content
        relevance_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Cosine similarity or RRF score
          example: 0.89
        ranking_position:
          type: integer
          minimum: 1
          description: Result rank (1 = top result)
          example: 1
        source_document:
          type: string
          description: Document name
          example: Safety_Manual_2024.pdf
        chapter_context:
          type: object
          properties:
            chapter_number:
              type: string
              example: "5.2.1"
            chapter_title:
              type: string
              example: Emergency Procedures
            chapter_path:
              type: string
              example: "5 > 5.2 > 5.2.1"
        quality_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Chunk quality score
          example: 92.5
        metadata:
          type: object
          description: Additional chunk metadata
          properties:
            department:
              type: string
            category:
              type: string
            fleet_type:
              type: string
            chunk_index:
              type: integer
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: ValidationError
        message:
          type: string
          description: Human-readable error message
          example: Query text must be at least 3 characters
        details:
          type: object
          description: Additional error details